requirements.txt
pytest==8.3.2
pytest-xdist==3.6.1
playwright==1.46.0
requests==2.32.3

allure-pytest==2.13.5

python-dotenv==1.0.1
pydantic==2.8.2

ruff==0.6.8
black==24.8.0
pre-commit==3.8.0
